pyproject.toml
[project]
name = "local-image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "daft>=0.7.2",
    "huggingface-hub>=0.36.0",
    "matplotlib>=3.10.8",
    "mlx>=0.30.3",
    "mlx-data>=0.2.0",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "pylance>=1.0.2",
    "torch>=2.9.1",
    "transformers>=4.57.6",
]
